Pulled out test for raw Antlr
