+++ src/test/scala/com/atomist/rug/kind/python3/Python3ParserTest.scala
@@ -1,19 +1,10 @@
 package com.atomist.rug.kind.python3
 
-import com.atomist.parse.java.ParsingTargets
-import com.atomist.project.archive.DefaultAtomistConfig
-import com.atomist.rug.kind.DefaultTypeRegistry
-import com.atomist.rug.kind.core.ProjectMutableView
-import com.atomist.rug.kind.java.JavaClassOrInterfaceView
-import com.atomist.source.{EmptyArtifactSource, SimpleFileBasedArtifactSource, StringFileArtifact}
 import com.atomist.tree.content.text.MutableContainerTreeNode
-import com.atomist.tree.pathexpression.{PathExpressionEngine, PathExpressionParser}
-import com.atomist.tree.utils.TreeNodeUtils
 import org.scalatest.{FlatSpec, Matchers}
 
-class Python3ParserTest extends FlatSpec with Matchers {
 
-  private val pex = new PathExpressionEngine
+class Python3ParserTest extends FlatSpec with Matchers {
 
   lazy val parser = new Python3Parser
 
@@ -44,30 +35,6 @@ class Python3ParserTest extends FlatSpec with Matchers {
 
   it should "parse simple script and write out unchanged" in
     parseAndVerifyValueCanBeWrittenOutUnchanged(dateParserDotPy)
-
-  it should "find Python file type using path expression" in {
-    val proj = SimpleFileBasedArtifactSource(StringFileArtifact("src/setup.py", setupDotPy))
-    val pmv = new ProjectMutableView(EmptyArtifactSource(""), proj, DefaultAtomistConfig)
-    val expr = "/src/File()/PythonFile()"
-    val rtn = pex.evaluate(pmv, PathExpressionParser.parseString(expr), DefaultTypeRegistry)
-    rtn.right.get.size should be(1)
-//    rtn.right.get.foreach {
-//      case p: PythonFileMutableView =>
-//    }
-  }
-
-  it should "drill down to Python import statement using path expression" in {
-    val proj = SimpleFileBasedArtifactSource(StringFileArtifact("src/setup.py", setupDotPy))
-    val pmv = new ProjectMutableView(EmptyArtifactSource(""), proj, DefaultAtomistConfig)
-    val expr = "/src/File()/PythonRawFile()//import_stmt()"
-    val rtn = pex.evaluate(pmv, PathExpressionParser.parseString(expr), DefaultTypeRegistry)
-    rtn.right.get.size should be>(2)
-    rtn.right.get.foreach {
-      case n: MutableContainerTreeNode =>
-        println(n.value)
-      case x => println(x)
-    }
-  }
 
   private def parseAndVerifyValueCanBeWrittenOutUnchanged(pyProg: String): MutableContainerTreeNode = {
     if (pyProg.lines.exists(_.trim.startsWith("|")))
